change completion expected and make optional
completion expected more than one hour seems to be causing an issue on the my dashboard with duplicate entries for zoom meetings, so changing it to 5 minutes after meeting start universally. Adding an additional setting to turn off completion expected entirely, in case the 5 minute setting does not resolve the issue.